@@ -4,4 +4,4 @@
 # If you change this key, all old signed cookies will become invalid!
 # Make sure the secret is at least 30 characters and all random,
 # no regular words or you'll be exposed to dictionary attacks.
-SoundCloudCompetition::Application.config.secret_token = ''
+SoundCloudCompetition::Application.config.secret_token = '4e73f4cd94317cf96ecc2d76b88d69b0e8b75af336f7a1f2a6a89a413002b4f9550d54958cf444b759a3ca0de58c682b2f677590eca512dcf6f41cea0d3906bc'
